refactor: move function to a common file
