+++ src/entity/ExperimentVariant.ts
@@ -1,5 +1,4 @@
 import { Column, Entity, PrimaryColumn } from 'typeorm';
-import { ConnectionManager } from './posts';
 
 @Entity()
 export class ExperimentVariant {
@@ -15,10 +14,3 @@ export class ExperimentVariant {
   @Column({ type: 'text', default: null })
   value: string;
 }
-
-export const getExperimentVariant = (
-  con: ConnectionManager,
-  feature: string,
-  variant: string,
-) =>
-  con.getRepository(ExperimentVariant).findOne({ where: { feature, variant } });
